Broadcast delay tables to the correlator nodes using MPI_Bcast, rather than point to point transfer of the delay tables from the manager node to each correlator node.
In multiple phase center mode we compute uvw and delay for each source
in the scan. Transferring all these delays point to point can take up 
to several minutes which is not acceptable for eVLBI. This is solved 
by using MPI_Bcast instead which uses a tree algorithm to greatly 
optimize the data transfer.