Merge fix to data_reader_file to stable branch
------------------------------------------------------------------------
r1787 | keimpema | 2014-09-02 18:22:14 +0200 (Tue, 02 Sep 2014) | 6 lines

When the data_reader_file encountered a bad file, e.g. due to a broken disk
it would abort the correlation. This behaviour caused correlation on the
Flexbuff to fail. The data_reader_file will now just print an error message
to the log and move to the next file when it can't open a file.